Create Tile Mask Package in raster Project
