+++ raster/src/main/scala/geotrellis/raster/mask/Implicits.scala
@@ -1,0 +1,9 @@
+package geotrellis.raster.mask
+
+import geotrellis.raster._
+
+object Implicits extends Implicits
+
+trait Implicits {
+  implicit class withTileMaskMethods(val self: Tile) extends TileMaskMethods
+}
